refactor: modularize CollectionAllocErr into errors.rs
Move CollectionAllocErr enum and its Display/Error impls from lib.rs
to a dedicated src/errors.rs module, re-exporting it from lib.rs to
preserve the public API.

Fixes #513
